@@ -12,44 +12,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/platform-server": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "@ngrx/effects": "^4.1.1",
-    "@ngrx/router-store": "^4.1.1",
-    "@ngrx/store": "^4.1.1",
-    "@ngrx/store-devtools": "^4.1.1",
+    "@angular/animations": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/platform-server": "6.0.2",
+    "@angular/router": "6.0.2",
+    "@ngrx/effects": "^6.0.0-beta.1",
+    "@ngrx/router-store": "^6.0.0-beta.1",
+    "@ngrx/store": "^6.0.0-beta.1",
+    "@ngrx/store-devtools": "^6.0.0-beta.1",
     "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
-    "firebase": "^3.7.0",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.5"
+    "firebase": "^4.6.0",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.0",
-    "@angular/compiler-cli": "4.0.0",
+    "@angular/cli": "^6.0.3",
+    "@angular/compiler-cli": "6.0.2",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
+    "codelyzer": "^2.0.1",
     "jasmine-core": "~2.5.2",
     "jasmine-spec-reporter": "~3.2.0",
     "karma": "~1.4.1",
     "karma-chrome-launcher": "~2.0.0",
     "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.4.2",
-    "typescript": "~2.4.2"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.3"
   }
 }
