@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { FormGroup, FormControl, FormArray, Validators } from '@angular/forms';
 import { Store } from '@ngrx/store';
-import 'rxjs/add/operator/take';
+import { take } from 'rxjs/operators';
 
 import * as RecipeActions from '../store/recipe.actions';
 import * as fromRecipe from '../store/recipe.reducers';
@@ -74,7 +74,7 @@ export class RecipeEditComponent implements OnInit {
 
     if (this.editMode) {
       this.store.select('recipes')
-        .take(1)
+        .pipe(take(1))
         .subscribe((recipeState: fromRecipe.State) => {
           const  recipe = recipeState.recipes[this.id];
           recipeName = recipe.name;
