@@ -2,8 +2,7 @@ import { Injectable } from '@angular/core';
 import { Actions, Effect } from '@ngrx/effects';
 import { HttpClient, HttpRequest } from '@angular/common/http';
 import { Store } from '@ngrx/store';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/withLatestFrom';
+import { switchMap, withLatestFrom, map } from 'rxjs/operators';
 
 import * as RecipeActions from './recipe.actions';
 import * as fromRecipe from '../../recipes/store/recipe.reducers';
@@ -15,15 +14,14 @@ export class RecipeEffects {
 
   @Effect()
   recipeFetch = this.actions$
-    .ofType(RecipeActions.FETCH_RECIPES)
-    .switchMap((action: RecipeActions.FetchRecipes) => {
+    .ofType(RecipeActions.FETCH_RECIPES).pipe(
+      switchMap((action: RecipeActions.FetchRecipes) => {
       return this.http.get<Recipe[]>(this.URL,
         {
-        observe: 'body', // default
-        responseType: 'json', // default
-      });
-    })
-    .map(
+          observe: 'body', // default
+          responseType: 'json', // default
+        });
+    }), map(
       (recipes) => {
         for (const recipe of recipes) {
           if (!recipe['ingredients']) {
@@ -34,17 +32,21 @@ export class RecipeEffects {
           type: RecipeActions.SET_RECIPES,
           payload: recipes
         };
-    });
+      })
+    );
+
 
   @Effect({dispatch: false})
   resipeStore = this.actions$
     .ofType(RecipeActions.STORE_RECIPES)
-    .withLatestFrom(this.store.select('recipes'))
-    .switchMap(([action, state]) => {
+    .pipe(
+    withLatestFrom(this.store.select('recipes')),
+    switchMap(([action, state]) => {
       const req = new HttpRequest('PUT', this.URL,
         state.recipes, {reportProgress: true});
       return this.http.request(req);
-    });
+    })
+    );
 
   constructor(private actions$: Actions,
               private http: HttpClient,
