@@ -1,6 +1,7 @@
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { Injectable } from '@angular/core';
+import { take, map } from 'rxjs/operators';
 
 import * as fromApp from '../store/app.reducers';
 import * as fromAuth from './store/auth.reducers';
@@ -12,9 +13,10 @@ export class AuthGuard implements CanActivate {
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
     return this.store.select('auth')
-      .take(1)
-      .map((authState: fromAuth.State) => {
-      return authState.authenticated;
-    });
+      .pipe(take(1),
+        map((authState: fromAuth.State) => {
+          return authState.authenticated;
+        })
+      );
   }
 }
